# frontend/openapi/wst-openapi.yaml
openapi: 3.1.0
info:
  title: Workshop Management & Student Practical Training (WST) API
  version: 1.0.0
  summary: Contract-first REST API for the WST academic MVP.
  description: |
    Source of truth for the PostgreSQL schema, Node.js backend, automated tests, React client and documentation.
    Requirement prefix: WST. Every operation carries `x-requirement-id` and `x-priority`.

    ## Conventions
    - Base path `/api/v1`. JSON bodies (multipart only for attachment upload). Bearer JWT. Operations with `security: []` are public.
    - Identifiers are UUIDs. Human-readable numbers (`jobNumber`, `poNumber`, `invoiceNumber`, `certificateNumber`) are server-generated, unique, immutable, and never localized or reordered by the backend.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Money is `{amount, currency}`: `amount` is a decimal string (never a float) and `currency` an ISO-4217 code. One system currency is configured in finance settings; documents in another currency are rejected with 422. Part quantities are whole numbers. Rates and metric values are decimal strings.
    - Optional properties without a value are omitted, never null. `readOnly` properties are server-controlled. Request schemas forbid unknown properties, so clients cannot submit ids, audit fields, stock balances, invoice totals, or permissions.
...

    ## Collections
...

    ## Authorization
...

    ## Concurrency and idempotency
...

    ## Localization
...

    ## Errors and rate limiting
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      operationId: listCustomers
      summary: List customers
      description: "Requires `customers.read`. Rows limited to the caller's organization scopes. Sortable: displayName, createdAt (default displayName)."
      x-requirement-id: [WST-FR-03, WST-FR-01]
      x-priority: Must
      x-permissions: [customers.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, ARCHIVED] } }
        - { name: phone, in: query, schema: { type: string, maxLength: 20 } }
      responses:
        '200': { description: Customers, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Customers]
      operationId: createCustomer
      summary: Register a customer
      description: "Requires `customers.write`. Customers are business entities only; there is no customer login."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [customers.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /customers/{customerId}:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      tags: [Customers]
      operationId: getCustomer
      summary: Get a customer
      description: "Requires `customers.read`. Contact data is sensitive; reads by staff are not individually audited but bulk exports are."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [customers.read]
      responses:
        '200': { description: Customer, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Customers]
      operationId: updateCustomer
      summary: Update or archive a customer
      description: "Requires `customers.write`. Setting status ARCHIVED is rejected with 409 RESOURCE_IN_USE while the customer has non-delivered jobs."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [customers.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /customers/{customerId}/statement:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      tags: [Invoices]
      operationId: getCustomerStatement
      summary: Customer invoice statement
      description: "Requires `invoices.read`. Returns the customer display name only (no contact data). Use an ExportJob of type CUSTOMER_STATEMENT for a PDF or CSV."
      x-requirement-id: [WST-FR-09]
      x-priority: Should
      x-permissions: [invoices.read]
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200': { description: Statement, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerStatement' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /bays:
    get:
      tags: [Bays]
      operationId: listBays
      summary: List workshop bays
      description: "Requires `bays.read`. Bays are shared by workshop jobs and training sessions. Sortable: code, name."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [bays.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, MAINTENANCE, INACTIVE] } }
      responses:
        '200': { description: Bays, content: { application/json: { schema: { $ref: '#/components/schemas/BayPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Bays]
      operationId: createBay
      summary: Create a bay
      description: "Requires `bays.manage`."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [bays.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BayCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Bay' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /bays/{bayId}:
    parameters:
      - $ref: '#/components/parameters/BayId'
    patch:
      tags: [Bays]
      operationId: updateBay
      summary: Update a bay or change its availability status
      description: "Requires `bays.manage`. Setting MAINTENANCE or INACTIVE does not cancel existing bookings; it makes future publishing or assignment fail with BAY_UNAVAILABLE."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [bays.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BayUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Bay' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /bays/{bayId}/calendar:
    parameters:
      - $ref: '#/components/parameters/BayId'
    get:
      tags: [Bays]
      operationId: getBayCalendar
      summary: Shared bay calendar (jobs and training sessions)
      description: "Requires `bays.read`. One calendar for both workshop jobs (RECEIVED, IN_PROGRESS, QUALITY_CHECK with an assigned window) and training sessions (DRAFT excluded, PUBLISHED and COMPLETED included). Job entries expose only the job number, never customer or vehicle data. Maximum window 31 days."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [bays.read]
      parameters:
        - { name: from, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
        - { name: to, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
      responses:
        '200': { description: Calendar, content: { application/json: { schema: { $ref: '#/components/schemas/BayCalendar' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /job-cards/{jobId}/part-issues:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Inventory]
      operationId: listPartIssues
      summary: List parts issued to a job
      description: "Requires `inventory.read`, `inventory.issue` or `inventory.reverse`, plus organization scope (jobs.read is not required); technicians see only assigned jobs. Cost fields need `inventory.cost.read`. Sortable: createdAt."
      x-requirement-id: [WST-FR-06]
      x-priority: Must
      x-permissions: [inventory.read, inventory.issue, inventory.reverse]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': { description: Part issues, content: { application/json: { schema: { $ref: '#/components/schemas/PartIssuePage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Inventory]
      operationId: issuePart
      summary: Issue a part to a job and deduct stock transactionally
      description: "Requires `inventory.issue` and assignment to the job. Job must be IN_PROGRESS (JOB_STAGE_NOT_ALLOWED) with an APPROVED INITIAL_WORK approval, and the approval of any referenced additional work item, else 409 CUSTOMER_APPROVAL_REQUIRED. In one database transaction the balance row is locked, quantity is checked against available stock (on-hand minus reserved, plus the caller's own reservation), an ISSUE ledger movement is appended, the balance is updated and the PartIssue snapshots unit price and average cost. Concurrent issues serialize on the balance so stock never goes negative; a shortfall returns 409 INSUFFICIENT_STOCK. After commit the reorder rule may create an advisory suggestion; it never creates a purchase order."
      x-requirement-id: [WST-FR-06, WST-FR-07, WST-FR-14]
      x-priority: Must
      x-permissions: [inventory.issue]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartIssueCreateRequest' }
            examples:
              issue: { $ref: '#/components/examples/PartIssueExample' }
      responses:
        '201':
          description: Issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartIssue' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Approval missing, stage not allowed, insufficient stock or idempotency conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                insufficientStock: { $ref: '#/components/examples/InsufficientStockExample' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/part-issues/{partIssueId}/reversals:
    parameters:
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/PartIssueId'
    post:
      tags: [Inventory]
      operationId: reversePartIssue
      summary: Reverse all or part of a part issue
      description: "Requires `inventory.reverse`. Reason is mandatory. quantity may not exceed issued minus already reversed (409 REVERSAL_EXCEEDS_ISSUED). Allowed while the job is IN_PROGRESS or QUALITY_CHECK. One transaction appends an ISSUE_REVERSAL movement that restores on-hand quantity, updates the issue status and writes an audit event. The original issue and movement are never modified."
      x-requirement-id: [WST-FR-06, WST-FR-07]
      x-priority: Must
      x-permissions: [inventory.reverse]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PartIssueReversalRequest' } } } }
      responses:
        '201': { description: Reversed, content: { application/json: { schema: { $ref: '#/components/schemas/PartIssueReversal' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token issued by login or refresh. The token identifies the user only; roles, permissions and scopes are resolved server-side on each request."
  parameters:
    Page: { name: page, in: query, description: "1-based page number.", schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, description: "Items per page, maximum 100.", schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, description: "Comma-separated fields; prefix - for descending. Only fields named by the operation are accepted.", schema: { type: string, maxLength: 100, pattern: '^-?[A-Za-z][A-Za-z0-9]*(,-?[A-Za-z][A-Za-z0-9]*)*$' } }
    From: { name: from, in: query, description: "Inclusive lower bound (RFC 3339 UTC).", schema: { $ref: '#/components/schemas/Timestamp' } }
    To: { name: to, in: query, description: "Exclusive upper bound (RFC 3339 UTC).", schema: { $ref: '#/components/schemas/Timestamp' } }
    Query: { name: q, in: query, description: "Case-insensitive text search.", schema: { type: string, minLength: 1, maxLength: 100 } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, description: "Client-generated key that makes the POST safely repeatable.", schema: { type: string, minLength: 8, maxLength: 128 } }
...
    CustomerId: { name: customerId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
    BayId: { name: bayId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    JobId: { name: jobId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
    PartIssueId: { name: partIssueId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    BadRequest:
      description: "Malformed request, unknown filter or sort field, or invalid query parameter."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: "Missing, invalid or expired credentials."
      headers: { WWW-Authenticate: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: "Authenticated but lacking the required permission. Rows outside the caller's scope are reported as 404 instead."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: "Resource does not exist or is outside the caller's row scope."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: "Invalid state transition, missing customer approval, insufficient stock, concurrency conflict, separation-of-duties violation or duplicate business operation. See Error.code."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ValidationError:
      description: "Semantic validation failed; details identify each offending field."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  examples:
...
    PartIssueExample:
      summary: Issue two units to a job
      value: { partId: 7c0d2f60-5a1b-4f0e-8d3c-2b9e4a1f6c77, storeId: 1a9b8c7d-6e5f-4a3b-9c2d-1e0f9a8b7c6d, quantity: 2 }
    InsufficientStockExample:
      summary: Concurrent issue exhausted stock
      value: { code: INSUFFICIENT_STOCK, message: "Requested quantity exceeds available stock.", details: [ { field: /quantity, code: INSUFFICIENT_STOCK, message: "Only 1 unit available.", params: { available: '1', requested: '2' } } ], requestId: 01J9ZK3Q8W2V5X7Y1B4N6M8P0R }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time, description: "RFC 3339 UTC timestamp." }
...
    CurrencyCode: { type: string, pattern: '^[A-Z]{3}$', description: "ISO-4217 alphabetic currency code." }
    MoneyAmount: { type: string, pattern: '^-?[0-9]{1,12}(\.[0-9]{1,4})?$', description: "Decimal string; never a JSON number." }
    Money:
      type: object
      required: [amount, currency]
      properties: { amount: { $ref: '#/components/schemas/MoneyAmount' }, currency: { $ref: '#/components/schemas/CurrencyCode' } }
...
    Locale: { type: string, enum: [en, ar] }
...
    RecordMeta:
      type: object
      required: [id, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: '#/components/schemas/Uuid', readOnly: true }
        createdAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        updatedAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        createdBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
        updatedBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
    PageInfo:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1, maximum: 100 }, totalItems: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ErrorCode:
      type: string
      enum: [BAD_REQUEST, UNAUTHENTICATED, INVALID_CREDENTIALS, TOKEN_EXPIRED, FORBIDDEN, NOT_FOUND, VALIDATION_FAILED, RATE_LIMITED, INTERNAL_ERROR, SERVICE_UNAVAILABLE, INVALID_STATE_TRANSITION, VERSION_CONFLICT, DUPLICATE_RESOURCE, IDEMPOTENCY_CONFLICT, RESOURCE_IN_USE, CUSTOMER_APPROVAL_REQUIRED, JOB_STAGE_NOT_ALLOWED, JOB_ASSIGNMENT_REQUIRED, CHECKLIST_INCOMPLETE, QUALITY_CHECK_REQUIRED, INVOICE_REQUIRED, INSUFFICIENT_STOCK, REVERSAL_EXCEEDS_ISSUED, RESERVATION_INVALID, SCHEDULE_CONFLICT, CONFLICT_NOT_OVERRIDABLE, SEPARATION_OF_DUTIES_VIOLATION, DUPLICATE_APPROVAL, RECEIPT_EXCEEDS_ORDERED, PAYMENT_AMOUNT_MISMATCH, ASSESSMENT_LOCKED, CERTIFICATE_NOT_ELIGIBLE, ATTACHMENT_INVALID, ATTACHMENT_NOT_LINKABLE, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, EXPORT_NOT_READY, EXPORT_EXPIRED]
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        field: { type: string, description: "JSON Pointer to the offending request field." }
        code: { type: string, description: "Machine-readable detail code, for example REQUIRED, TOO_LONG, MILEAGE_LOWER_THAN_RECORDED." }
        message: { type: string }
        params: { type: object, additionalProperties: { type: string }, description: "Named values for message templates, for example available and requested quantities." }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, description: "Human-readable message in the negotiated language." }
        details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        requestId: { type: string, description: "Correlation id also present in logs and audit events." }
...
    ContactPreferences:
      type: object
      properties: { preferredChannel: { type: string, enum: [PHONE, SMS, EMAIL] }, preferredLocale: { $ref: '#/components/schemas/Locale' } }
      additionalProperties: false
    Customer:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [organizationScopeId, displayName, type, phone, status]
          properties:
            organizationScopeId: { $ref: '#/components/schemas/Uuid' }
            displayName: { type: string, minLength: 1, maxLength: 160 }
            type: { type: string, enum: [INDIVIDUAL, BUSINESS] }
            phone: { type: string, pattern: '^\+[1-9][0-9]{6,14}$' }
            email: { type: string, format: email }
            contactPreferences: { $ref: '#/components/schemas/ContactPreferences' }
            status: { type: string, enum: [ACTIVE, ARCHIVED] }
    CustomerCreateRequest:
      type: object
      required: [organizationScopeId, displayName, type, phone]
      additionalProperties: false
      properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, displayName: { type: string, minLength: 1, maxLength: 160 }, type: { type: string, enum: [INDIVIDUAL, BUSINESS] }, phone: { type: string, pattern: '^\+[1-9][0-9]{6,14}$' }, email: { type: string, format: email }, contactPreferences: { $ref: '#/components/schemas/ContactPreferences' } }
    CustomerUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { displayName: { type: string, minLength: 1, maxLength: 160 }, phone: { type: string, pattern: '^\+[1-9][0-9]{6,14}$' }, email: { type: string, format: email }, contactPreferences: { $ref: '#/components/schemas/ContactPreferences' }, status: { type: string, enum: [ACTIVE, ARCHIVED] } }
...
    Bay:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [organizationScopeId, code, name, capacity, status]
          properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, code: { type: string, minLength: 1, maxLength: 20 }, name: { type: string, maxLength: 80 }, capacity: { type: integer, minimum: 0, description: "Maximum training participants." }, status: { type: string, enum: [ACTIVE, MAINTENANCE, INACTIVE] } }
    BayCreateRequest:
      type: object
      required: [organizationScopeId, code, name, capacity]
      additionalProperties: false
      properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, code: { type: string, minLength: 1, maxLength: 20 }, name: { type: string, maxLength: 80 }, capacity: { type: integer, minimum: 0 } }
    BayUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { name: { type: string, maxLength: 80 }, capacity: { type: integer, minimum: 0 }, status: { type: string, enum: [ACTIVE, MAINTENANCE, INACTIVE] } }
    CalendarReference:
      type: object
      required: [kind, referenceId, referenceLabel, startsAt, endsAt]
      properties: { kind: { type: string, enum: [JOB, TRAINING_SESSION] }, referenceId: { $ref: '#/components/schemas/Uuid' }, referenceLabel: { type: string, description: "Job number or session title; never customer or vehicle data." }, startsAt: { $ref: '#/components/schemas/Timestamp' }, endsAt: { $ref: '#/components/schemas/Timestamp' } }
    BayCalendar:
      type: object
      required: [bayId, from, to, entries]
      properties: { bayId: { $ref: '#/components/schemas/Uuid' }, from: { $ref: '#/components/schemas/Timestamp' }, to: { $ref: '#/components/schemas/Timestamp' }, entries: { type: array, items: { $ref: '#/components/schemas/CalendarReference' } } }
    JobNumber: { type: string, pattern: '^JC-[0-9]{4}-[0-9]{6}$', description: "Server-generated, unique, human-readable. Always rendered LTR by clients." }
...
    PartIssue:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [jobId, partId, partSku, storeId, quantity, reversedQuantity, unitPrice, lineTotal, status, stockMovementId]
          properties: { jobId: { $ref: '#/components/schemas/Uuid' }, partId: { $ref: '#/components/schemas/Uuid' }, partSku: { type: string }, storeId: { $ref: '#/components/schemas/Uuid' }, workItemId: { $ref: '#/components/schemas/Uuid' }, reservationId: { $ref: '#/components/schemas/Uuid' }, quantity: { type: integer, minimum: 1 }, reversedQuantity: { type: integer, minimum: 0 }, unitPrice: { $ref: '#/components/schemas/Money' }, unitCost: { $ref: '#/components/schemas/Money', description: "Only with inventory.cost.read." }, lineTotal: { $ref: '#/components/schemas/Money' }, status: { type: string, enum: [ISSUED, PARTIALLY_REVERSED, REVERSED] }, stockMovementId: { $ref: '#/components/schemas/Uuid' } }
    PartIssueCreateRequest:
      type: object
      required: [partId, storeId, quantity]
      additionalProperties: false
      properties: { partId: { $ref: '#/components/schemas/Uuid' }, storeId: { $ref: '#/components/schemas/Uuid' }, quantity: { type: integer, minimum: 1, maximum: 100000 }, workItemId: { $ref: '#/components/schemas/Uuid' }, reservationId: { $ref: '#/components/schemas/Uuid' } }
    PartIssueReversalRequest:
      type: object
      required: [quantity, reason]
      additionalProperties: false
      properties: { quantity: { type: integer, minimum: 1 }, reason: { type: string, minLength: 3, maxLength: 500 } }
    PartIssueReversal:
      type: object
      required: [id, partIssueId, quantity, reason, reversedBy, reversedAt, stockMovementId, partIssueStatus]
      properties: { id: { $ref: '#/components/schemas/Uuid' }, partIssueId: { $ref: '#/components/schemas/Uuid' }, quantity: { type: integer, minimum: 1 }, reason: { type: string }, reversedBy: { $ref: '#/components/schemas/Uuid' }, reversedAt: { $ref: '#/components/schemas/Timestamp' }, stockMovementId: { $ref: '#/components/schemas/Uuid' }, partIssueStatus: { type: string, enum: [ISSUED, PARTIALLY_REVERSED, REVERSED] } }
...
    InvoiceStatus: { type: string, enum: [DRAFT, ISSUED, PAID, VOID] }
...
    StatementLine:
      type: object
      required: [invoiceId, jobNumber, status, total, paidAmount]
      properties: { invoiceId: { $ref: '#/components/schemas/Uuid' }, invoiceNumber: { type: string }, jobNumber: { $ref: '#/components/schemas/JobNumber' }, issuedAt: { $ref: '#/components/schemas/Timestamp' }, status: { $ref: '#/components/schemas/InvoiceStatus' }, total: { $ref: '#/components/schemas/Money' }, paidAmount: { $ref: '#/components/schemas/Money' } }
    CustomerStatement:
      type: object
      required: [customerId, customerDisplayName, currencyCode, lines, totalInvoiced, totalPaid, totalOutstanding, generatedAt]
      properties: { customerId: { $ref: '#/components/schemas/Uuid' }, customerDisplayName: { type: string }, from: { $ref: '#/components/schemas/Timestamp' }, to: { $ref: '#/components/schemas/Timestamp' }, currencyCode: { $ref: '#/components/schemas/CurrencyCode' }, lines: { type: array, items: { $ref: '#/components/schemas/StatementLine' } }, totalInvoiced: { $ref: '#/components/schemas/Money' }, totalPaid: { $ref: '#/components/schemas/Money' }, totalOutstanding: { $ref: '#/components/schemas/Money' }, generatedAt: { $ref: '#/components/schemas/Timestamp' } }
...
    CustomerPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/Customer' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...
    BayPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/Bay' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...
    PartIssuePage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/PartIssue' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...